requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
httpx==0.25.0
python-dotenv==1.0.0
pydantic==2.3.0
python-multipart==0.0.6
openai==1.3.0
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
frozenlist==1.4.0
multidict==6.0.4
yarl==1.9.2
pydantic-settings==2.0.3